pyproject.toml
[project]
name = "datalink"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mkdocs-macros-plugin>=1.3.9",
    "mkdocs-material>=9.6.20",
    "mkdocs-gen-files>=0.5.0",
    "pyyaml>=6.0",
]
